requirements.txt
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.0.0
statsmodels>=0.13.0
nltk>=3.6.0
textblob>=0.15.3
wordcloud>=1.8.0
spacy>=3.0.0
streamlit>=1.10.0
plotly>=5.3.0
scipy>=1.7.0
openai>=0.27.0
langchain>=0.0.200
python-dotenv>=0.19.0
requests>=2.26.0
beautifulsoup4>=4.9.0
openpyxl>=3.0.0
xlrd>=2.0.0
PyYAML>=6.0.0
tqdm>=4.62.0 